Handle extraneous carriage returns in data
A couple of these appeared in a `GS_CASE` file I was loading, and
`fixed2csv` treated them as if they weren't there at all--which threw
off all of the subsequent fields in that row.

These characters now are replaced with normal spaces, which might or
might not be semantically correct for a given field but at least won't
wreck everything to the right of it.