Use partitions to keep separate files separate
If DOJ provides us with a table that's been divided up by judicial
district, let's go ahead and keep it divided that way.

Ideally, we'd remove the `district` column entirely from the spec and
use this instead, rather than treat it as a completely separate
`filename_district` column, but this is OK for a first pass.