Force a four-digit year in all output
We encountered a mistyped date in the NCD: `11-MAR-0024` (presumably
intended to refer to March 11, 2024, rather than that date in the actual
year A.D. 24).

On the Mac where I've been developing, the `strptime`-to-`strftime`
conversion outputs this as `0024-03-11`, which is what both Postgres and
I would expect. On the Ubuntu 18.04 machine where I just attempted to run
the import script, it instead output `24-03-11`, and Postgres rightfully
raised an exception out of concern that the date format was inconsistent.

I'd rather propagate the original erroneous year value of `0024`, so we
now have this `rjust` call in order to enforce the database's requirement
of a four-digit year.